Add files via upload
These files are for the Google Capstone Project. The premise of this project was to aid the marketing team of a ride-share bicycle company. My goal for this project was to answer how users ride differently to identify how the marketing team should proceed. The dataset came from https://divvy-tripdata.s3.amazonaws.com/index.html which is made available to the public. License: https://www.divvybikes.com/data-license-agreement Thank you for your time